+++ src/adapter/web-standard/handler.ts
@@ -4,12 +4,10 @@ import {
 	createResponseHandler,
 	createStreamHandler,
 	handleFile,
-	handleSet,
-	responseToSetHeaders,
-	streamResponse
+	handleSet
 } from '../utils'
 
-import { ElysiaFile } from '../../universal/file'
+import { ElysiaFile, mime } from '../../universal/file'
 import { isNotEmpty } from '../../utils'
 import { Cookie } from '../../cookies'
 import { ElysiaCustomStatusResponse } from '../../error'
@@ -24,8 +22,9 @@ const handleElysiaFile = (
 	}
 ) => {
 	const path = file.path
-	// @ts-ignore
-	const contentType = mime[path.slice(path.lastIndexOf('.') + 1)]
+	const contentType =
+		mime[path.slice(path.lastIndexOf('.') + 1) as any as keyof typeof mime]
+
 	if (contentType) set.headers['content-type'] = contentType
 
 	if (
@@ -85,30 +84,6 @@ export const mapResponse = (
 					request
 				)
 
-			case 'ReadableStream':
-				if (
-					!set.headers['content-type']?.startsWith(
-						'text/event-stream'
-					)
-				)
-					set.headers['content-type'] =
-						'text/event-stream; charset=utf-8'
-
-				request?.signal?.addEventListener(
-					'abort',
-					{
-						handleEvent() {
-							if (request?.signal && !request?.signal?.aborted)
-								(response as ReadableStream).cancel()
-						}
-					},
-					{
-						once: true
-					}
-				)
-
-				return new Response(response as ReadableStream, set as any)
-
 			case undefined:
 				if (!response) return new Response('', set as any)
 
@@ -167,8 +142,11 @@ export const mapResponse = (
 					)
 				}
 
-				// @ts-expect-error
-				if (typeof response?.next === 'function')
+				if (
+					// @ts-expect-error
+					typeof response?.next === 'function' ||
+					response instanceof ReadableStream
+				)
 					return handleStream(response as any, set, request) as any
 
 				// @ts-expect-error
@@ -197,17 +175,6 @@ export const mapResponse = (
 				return new Response(response as any, set as any)
 		}
 	}
-
-	if (
-		response instanceof Response &&
-		!(response as Response).headers.has('content-length') &&
-		(response as Response).headers.get('transfer-encoding') === 'chunked'
-	)
-		return handleStream(
-			streamResponse(response),
-			responseToSetHeaders(response as Response, set),
-			request
-		) as any
 
 	// Stream response defers a 'set' API, assume that it may include 'set'
 	if (
@@ -257,30 +224,6 @@ export const mapEarlyResponse = (
 					set,
 					request
 				)
-
-			case 'ReadableStream':
-				if (
-					!set.headers['content-type']?.startsWith(
-						'text/event-stream'
-					)
-				)
-					set.headers['content-type'] =
-						'text/event-stream; charset=utf-8'
-
-				request?.signal?.addEventListener(
-					'abort',
-					{
-						handleEvent() {
-							if (request?.signal && !request?.signal?.aborted)
-								(response as ReadableStream).cancel()
-						}
-					},
-					{
-						once: true
-					}
-				)
-
-				return new Response(response as ReadableStream, set as any)
 
 			case undefined:
 				if (!response) return
@@ -340,8 +283,11 @@ export const mapEarlyResponse = (
 					)
 				}
 
-				// @ts-expect-error
-				if (typeof response?.next === 'function')
+				if (
+					// @ts-expect-error
+					typeof response?.next === 'function' ||
+					response instanceof ReadableStream
+				)
 					return handleStream(response as any, set, request) as any
 
 				// @ts-expect-error
@@ -398,26 +344,6 @@ export const mapEarlyResponse = (
 					request
 				)
 
-			case 'ReadableStream':
-				request?.signal?.addEventListener(
-					'abort',
-					{
-						handleEvent() {
-							if (request?.signal && !request?.signal?.aborted)
-								(response as ReadableStream).cancel()
-						}
-					},
-					{
-						once: true
-					}
-				)
-
-				return new Response(response as ReadableStream, {
-					headers: {
-						'Content-Type': 'text/event-stream; charset=utf-8'
-					}
-				})
-
 			case undefined:
 				if (!response) return new Response('')
 
@@ -428,17 +354,6 @@ export const mapEarlyResponse = (
 				})
 
 			case 'Response':
-				if (
-					!(response as Response).headers.has('content-length') &&
-					(response as Response).headers.get('transfer-encoding') ===
-						'chunked'
-				)
-					return handleStream(
-						streamResponse(response as Response),
-						responseToSetHeaders(response as Response),
-						request
-					) as any
-
 				return response as Response
 
 			case 'Promise':
@@ -488,8 +403,11 @@ export const mapEarlyResponse = (
 					)
 				}
 
-				// @ts-expect-error
-				if (typeof response?.next === 'function')
+				if (
+					// @ts-expect-error
+					typeof response?.next === 'function' ||
+					response instanceof ReadableStream
+				)
 					return handleStream(response as any, set, request) as any
 
 				// @ts-expect-error
@@ -557,26 +475,6 @@ export const mapCompactResponse = (
 				}
 			)
 
-		case 'ReadableStream':
-			request?.signal?.addEventListener(
-				'abort',
-				{
-					handleEvent() {
-						if (request?.signal && !request?.signal?.aborted)
-							(response as ReadableStream).cancel()
-					}
-				},
-				{
-					once: true
-				}
-			)
-
-			return new Response(response as ReadableStream, {
-				headers: {
-					'Content-Type': 'text/event-stream; charset=utf-8'
-				}
-			})
-
 		case undefined:
 			if (!response) return new Response('')
 
@@ -587,16 +485,6 @@ export const mapCompactResponse = (
 			})
 
 		case 'Response':
-			if (
-				(response as Response).headers.get('transfer-encoding') ===
-				'chunked'
-			)
-				return handleStream(
-					streamResponse(response as Response),
-					responseToSetHeaders(response as Response),
-					request
-				) as any
-
 			return response as Response
 
 		case 'Error':
@@ -639,8 +527,11 @@ export const mapCompactResponse = (
 					}
 				)
 
-			// @ts-expect-error
-			if (typeof response?.next === 'function')
+			if (
+				// @ts-expect-error
+				typeof response?.next === 'function' ||
+				response instanceof ReadableStream
+			)
 				return handleStream(response as any, undefined, request) as any
 
 			// @ts-expect-error
@@ -699,7 +590,7 @@ export const createStaticHandler = (
 		!hooks.beforeHandle?.length &&
 		!hooks.afterHandle?.length
 	)
-		return response.clone.bind(response) as any
+		return () => response.clone() as Response
 }
 
 const handleResponse = createResponseHandler({