+++ src/compose.ts
@@ -264,10 +264,14 @@ const composeValidationFactory = ({
 				!appliedCleaner && normalize && !noValidate
 
 			// Encode call TypeCheck.Check internally
-			if (encodeSchema && value.hasTransform)
+			if (encodeSchema && value.hasTransform) {
 				code +=
 					`try{` +
-					`${name}=validator.response[${status}].Encode(${name})\n` +
+					`${name}=validator.response[${status}].Encode(${name})\n`
+
+				if (!appliedCleaner) code += clean({ ignoreTryCatch: true })
+
+				code +=
 					`c.set.status=${status}` +
 					`}catch{` +
 					(applyErrorCleaner
@@ -279,7 +283,7 @@ const composeValidationFactory = ({
 							`}`
 						: `throw new ValidationError('response',validator.response[${status}],${name})`) +
 					`}`
-			else {
+			} else {
 				if (!appliedCleaner) code += clean()
 
 				if (!noValidate)
@@ -410,13 +414,19 @@ export const composeHandler = ({
 			headers: app.setHeaders ?? {}
 		})
 
+		const isResponse =
+			handler instanceof Response ||
+			// @ts-ignore If it's not instanceof Response, it might be a polyfill (only on Node)
+			(handler?.constructor?.name === 'Response' &&
+				typeof (handler as Response)?.clone === 'function')
+
 		if (
 			hooks.parse?.length &&
 			hooks.transform?.length &&
 			hooks.beforeHandle?.length &&
 			hooks.afterHandle?.length
 		) {
-			if (handler instanceof Response)
+			if (isResponse)
 				return Function(
 					'a',
 					'"use strict";\n' + `return function(){return a.clone()}`
@@ -426,6 +436,12 @@ export const composeHandler = ({
 				'a',
 				'"use strict";\n' + 'return function(){return a}'
 			)(handler)
+		}
+
+		if (isResponse) {
+			const response = handler as Response
+
+			handler = () => response.clone()
 		}
 	}
 
@@ -1915,8 +1931,6 @@ export const composeHandler = ({
 
 	fnLiteral = init + fnLiteral + '}'
 	init = ''
-
-	// console.log(fnLiteral)
 
 	try {
 		return Function(