+++ src/error.ts
@@ -248,6 +248,10 @@ export class ValidationError extends Error {
 	code = 'VALIDATION'
 	status = 422
 
+	valueError?: ValueError
+	expected?: unknown
+	customError?: string
+
 	constructor(
 		public type: string,
 		public validator: TSchema | TypeCheck<any> | ElysiaTypeCheck<any>,
@@ -262,30 +266,69 @@ export class ValidationError extends Error {
 			value = value.response
 
 		const error =
-			errors?.First() ||
-			(isProduction
-				? undefined
-				: 'Errors' in validator
-					? validator.Errors(value).First()
-					: Value.Errors(validator, value).First())
+			errors?.First() ??
+			('Errors' in validator
+				? validator.Errors(value).First()
+				: Value.Errors(validator, value).First())
+
+		const accessor = error?.path || 'root'
+
+		// @ts-ignore private field
+		const schema = validator?.schema ?? validator
+
+		let expected
+
+		if (!isProduction) {
+			try {
+				expected = Value.Create(schema)
+			} catch (error) {
+				expected = {
+					type: 'Could not create expected value',
+					// @ts-expect-error
+					message: error?.message,
+					error
+				}
+			}
+		}
 
 		const customError =
 			error?.schema?.message || error?.schema?.error !== undefined
 				? typeof error.schema.error === 'function'
-					? error.schema.error({
-							type,
-							validator,
-							value,
-							get errors() {
-								return [...validator.Errors(value)].map(
-									mapValueError
-								)
-							}
-						})
+					? error.schema.error(
+							isProduction
+								? {
+										type: 'validation',
+										on: type,
+										found: value
+									}
+								: {
+										type: 'validation',
+										on: type,
+										value,
+										property: accessor,
+										message: error?.message,
+										summary: mapValueError(error).summary,
+										found: value,
+										expected,
+										errors:
+											'Errors' in validator
+												? [
+														...validator.Errors(
+															value
+														)
+													].map(mapValueError)
+												: [
+														...Value.Errors(
+															validator,
+															value
+														)
+													].map(mapValueError)
+									},
+							validator
+						)
 					: error.schema.error
 				: undefined
 
-		const accessor = error?.path || 'root'
 		let message = ''
 
 		if (customError !== undefined) {
@@ -297,41 +340,24 @@ export class ValidationError extends Error {
 			message = JSON.stringify({
 				type: 'validation',
 				on: type,
-				summary: mapValueError(error).summary,
-				message: error?.message,
 				found: value
 			})
 		} else {
-			// @ts-ignore private field
-			const schema = validator?.schema ?? validator
-			const errors =
-				'Errors' in validator
-					? [...validator.Errors(value)].map(mapValueError)
-					: [...Value.Errors(validator, value)].map(mapValueError)
-
-			let expected
-
-			try {
-				expected = Value.Create(schema)
-			} catch (error) {
-				expected = {
-					type: 'Could not create expected value',
-					// @ts-expect-error
-					message: error?.message,
-					error
-				}
-			}
-
 			message = JSON.stringify(
 				{
 					type: 'validation',
 					on: type,
-					summary: mapValueError(error).summary,
 					property: accessor,
 					message: error?.message,
+					summary: mapValueError(error).summary,
 					expected,
 					found: value,
-					errors
+					errors:
+						'Errors' in validator
+							? [...validator.Errors(value)].map(mapValueError)
+							: [...Value.Errors(validator, value)].map(
+									mapValueError
+								)
 				},
 				null,
 				2
@@ -339,6 +365,10 @@ export class ValidationError extends Error {
 		}
 
 		super(message)
+
+		this.valueError = error
+		this.expected = expected
+		this.customError = customError
 
 		Object.setPrototypeOf(this, ValidationError.prototype)
 	}
@@ -375,5 +405,52 @@ export class ValidationError extends Error {
 				'content-type': 'application/json'
 			}
 		})
+	}
+
+	/**
+	 * Utility function to inherit add custom error and keep the original Validation error
+	 *
+	 * @since 1.3.14
+	 *
+	 * @example
+	 * ```ts
+	 * new Elysia()
+	 *		.onError(({ error, code }) => {
+	 *			if (code === 'VALIDATION') return error.detail(error.message)
+	 *		})
+	 *		.post('/', () => 'Hello World!', {
+	 *			body: t.Object({
+	 *				x: t.Number({
+	 *					error: 'x must be a number'
+	 *				})
+	 *			})
+	 *		})
+	 * ```
+	 */
+	detail(message: unknown) {
+		if (!this.customError) return this.message
+
+		const validator = this.validator
+		const value = this.value
+		const expected = this.expected
+		const errors = this.all
+
+		return isProduction
+			? {
+					type: 'validation',
+					on: this.type,
+					found: value,
+					message
+				}
+			: {
+					type: 'validation',
+					on: this.type,
+					property: this.valueError?.path || 'root',
+					message,
+					summary: mapValueError(this.valueError).summary,
+					found: value,
+					expected,
+					errors
+				}
 	}
 }