+++ src/type-system/utils.ts
@@ -10,7 +10,11 @@ import { TypeCheck, TypeCompiler } from '@sinclair/typebox/compiler'
 
 import { ElysiaFile } from '../universal/file'
 import { InvalidFileType, ValidationError } from '../error'
-import type { FileOptions, FileUnit } from './types'
+import type {
+	ElysiaTypeCustomErrorCallback,
+	FileOptions,
+	FileUnit
+} from './types'
 import type { MaybeArray } from '../types'
 
 export const tryParse = (v: unknown, schema: TAnySchema) => {
@@ -175,3 +179,28 @@ export const validateFile = (options: FileOptions, value: any) => {
 
 	return true
 }
+
+/**
+ * Utility function to inherit add custom error and keep the original Validation error
+ *
+ * @since 1.3.14
+ *
+ * @example
+ * ```ts
+ * import { Elysia, t, errorWithDetail } from 'elysia'
+ *
+ * new Elysia()
+ *		.post('/', () => 'Hello World!', {
+ *			body: t.Object({
+ *				x: t.Number({
+ *					error: validationDetail('x must be a number')
+ *				})
+ *			})
+ *		})
+ */
+export const validationDetail =
+	<T>(message: T) =>
+	(error: Parameters<ElysiaTypeCustomErrorCallback>[0]) => ({
+		...error,
+		message
+	})